memPlot.R formated; scj plotting support added

+++ oSCJ/examples/_bench/bench.sh
@@ -7,9 +7,8 @@
 #   --> plot in legend will have "name_of_this_run" as a name of that input
 #
 
-
-set -e
-set -x
+#set -e
+#set -x
 
 
 # ./bench.sh cdj_hf_A.1600 /Users/plsek/_work/fiji/fivm-scope/scj/oSCJ/examples/cdj-vanilla-fiji-v1.2/bench/regular_CDj_frame_1000_period_80_plane_60_GC_fifo10_detector_fifoMax
@@ -120,20 +119,22 @@ done
 if [ "$perf_files" == ""  ] ; then
     echo "Perf input is empty."
 else
-   ./R/performancePlot.R $perf_files
+   ./R/perfPlot.R $perf_files
 fi;
 
 
-#./R/performancePlot.R $perf_files
+#./R/perfPlot.R $perf_files
 #./R/heapPerf.R $perf_files
 
 
 #### SAVE PDF                                                                                                                                               
-FIGURES="perf_bench mem_bench WCET_vs_heap Time_vs_heap_box WCET_vs_trigger"
+FIGURES="perf_bench mem_bench Time_Distr_box WCET_vs_heap WCET_vs_trigger Memory_Distr_box"
 EXT="pdf"
 
 for fig in $FIGURES; do
-    cp $fig.$EXT ./plots/$regex$fig$timestamp.$EXT
+    if [ -e $fig.$EXT ]; then 
+	cp $fig.$EXT ./plots/$regex$fig$timestamp.$EXT
+    fi
 done
 
 rm -rf tmp
@@ -174,7 +175,7 @@ cp $FIJI/output_d.dat ./data/perf_.$timestamp_fiji.dat
 cp $FIJI/output_hg_level_a_d.dat ./data/fiji_mem_hg_level_a.$timestamp.dat 
 cp $FIJI/output_hg_100k_d.dat ./data/fiji_mem_hg_100k_a.$timestamp.dat 
 
-./R/performancePlot.R ./data/perf_$timestamp_scj.dat ./data/perf_.$timestamp_fiji.dat  ./data/fiji_mem_hg_level_a.$timestamp.dat ./data/fiji_mem_hg_level_a.$timestamp.dat #./data/fiji_mem_hg_100k_a.$timestamp.dat 
+./R/perfPlot.R ./data/perf_$timestamp_scj.dat ./data/perf_.$timestamp_fiji.dat  ./data/fiji_mem_hg_level_a.$timestamp.dat ./data/fiji_mem_hg_level_a.$timestamp.dat #./data/fiji_mem_hg_100k_a.$timestamp.dat 
 
 cp ./perf_bench.pdf ./plots/perf_bench.$timestamp.pdf
 #rm -rf perf_bench.pdf
